Fix bugs
- Fitness was too weak (branch count), replaced with state count
- Local search could diverge, weakened so that it stops after a while

+++ master/src/main/java/gasp/ga/operators/selection/SelectionFunctionRank.java
@@ -41,7 +41,7 @@ public class SelectionFunctionRank<T extends Gene<T>> implements SelectionFuncti
 	private int[] getRanks(List<Individual<T>> population) {
 		final int[] ranking = new int[population.size()];
 		int currRank = 0;
-        int currFitness = Integer.MIN_VALUE;
+        long currFitness = Long.MIN_VALUE;
         for (int i = population.size() - 1; i >= 0; --i) {
         	if (population.get(i).getFitness() > currFitness) {
         		currFitness = population.get(i).getFitness();
